Integrated project: Weaving the data threads of Maji Ndogo's narrative
I will pull data from many different tables and apply some statistical analyses to examine the consequences of an audit report that cross-references a random sample of records.
My objective is to assess the integrity, and accuracy of the data stored in the database. Auditing a database
involves verifying that the data it contains is both accurate and has not been tampered with, thereby ensuring that the information can be relied
upon for decision-making and governance.